turn to float

+++ LAB4.java
@@ -1,9 +1,9 @@
 public class LAB4 {
 	
-	public int add(int num1, int num2) {
+	public float add(int num1, int num2) {
 		return num1+num2;
 	}
-	public int minus(int num1, int num2) {
+	public float minus(int num1, int num2) {
 		return num1-num2;
 	}
 	public static void main(String[] args) {
